Address chat stream review cleanup
- Require Redis for stream status and constrain approval intervention types.\n- Share stream status typing and client fetch helper.\n- Simplify tool-use and interrupted-tool message typing.
