+++ web/src/lib/types/stream-status.ts
@@ -1,0 +1,6 @@
+export type StreamStatus = {
+    running: boolean
+    resumable: boolean
+    pendingApproval: boolean
+    pendingOAuth: boolean
+}
